Refactoring of error library
