+++ crates/cairoc/README.md
@@ -2,3 +2,4 @@
 
 ```suggestion
 A wrapper to generate a `FlatLowered` representation of Cairo code and convert it into the `flo` format to enable separate compilation and linking.
+```
