+++ crates/cli/src/main.rs
@@ -1,6 +1,5 @@
 //! This is the CLI driver for the compilation of LLVM IR to Cairo. For more
-//! detail, please see the documentation for the [`hieratika_compiler`] and
-//! [`hieratika_cairoc`] crate.
+//! detail, please see the documentation for the [`hieratika_cairoc`] crate.
 
 #![warn(clippy::all, clippy::cargo, clippy::pedantic)]
 #![allow(clippy::module_name_repetitions)] // Allows for better API naming
