+++ crates/compiler/src/lib.rs
@@ -50,7 +50,7 @@ pub mod llvm;
 pub mod pass;
 pub mod polyfill;
 
-use hieratika_errors::llvm_compile::{Error, Result};
+use hieratika_errors::compile::llvm::{Error, Result};
 use hieratika_flo::FlatLoweredObject;
 
 use crate::{
@@ -143,7 +143,7 @@ impl Compiler {
     ///
     /// # Errors
     ///
-    /// - [`hieratika_errors::llvm_compile::Error`] if the compilation process
+    /// - [`hieratika_errors::compile::llvm::Error`] if the compilation process
     ///   fails for any reason.
     pub fn run(mut self) -> Result<CompilationResult> {
         let PassManagerReturnData {
