+++ crates/compiler/src/llvm/data_layout.rs
@@ -6,7 +6,7 @@ use chumsky::{
     prelude::{choice, just},
     Parser,
 };
-use hieratika_errors::llvm_compile::{Error, Result};
+use hieratika_errors::compile::llvm::{Error, Result};
 
 use crate::constant::{
     BYTE_SIZE,
