+++ crates/compiler/src/llvm/typesystem.rs
@@ -3,7 +3,7 @@
 
 use std::fmt::{Display, Formatter};
 
-use hieratika_errors::{llvm_compile, llvm_compile::Error};
+use hieratika_errors::compile::{llvm, llvm::Error};
 use inkwell::types::{
     AnyTypeEnum,
     ArrayType,
@@ -274,7 +274,7 @@ impl Display for LLVMType {
 
 /// Conversion from Inkwell's generic type enum to our type language.
 impl<'ctx> TryFrom<AnyTypeEnum<'ctx>> for LLVMType {
-    type Error = llvm_compile::Error;
+    type Error = llvm::Error;
 
     fn try_from(value: AnyTypeEnum<'ctx>) -> Result<Self, Self::Error> {
         Self::try_from(&value)
@@ -283,7 +283,7 @@ impl<'ctx> TryFrom<AnyTypeEnum<'ctx>> for LLVMType {
 
 /// Conversion from Inkwell's generic type enum to our type language.
 impl<'ctx> TryFrom<&AnyTypeEnum<'ctx>> for LLVMType {
-    type Error = llvm_compile::Error;
+    type Error = llvm::Error;
 
     fn try_from(value: &AnyTypeEnum<'ctx>) -> Result<Self, Self::Error> {
         match value {
@@ -301,7 +301,7 @@ impl<'ctx> TryFrom<&AnyTypeEnum<'ctx>> for LLVMType {
 
 /// Conversion from Inkwell's basic type enum to our type language.
 impl<'ctx> TryFrom<BasicTypeEnum<'ctx>> for LLVMType {
-    type Error = llvm_compile::Error;
+    type Error = llvm::Error;
 
     fn try_from(value: BasicTypeEnum<'ctx>) -> Result<Self, Self::Error> {
         Self::try_from(&value)
@@ -310,7 +310,7 @@ impl<'ctx> TryFrom<BasicTypeEnum<'ctx>> for LLVMType {
 
 /// Conversion from Inkwell's basic type enum to our type language.
 impl<'ctx> TryFrom<&BasicTypeEnum<'ctx>> for LLVMType {
-    type Error = llvm_compile::Error;
+    type Error = llvm::Error;
 
     fn try_from(value: &BasicTypeEnum<'ctx>) -> Result<Self, Self::Error> {
         match value {
@@ -326,7 +326,7 @@ impl<'ctx> TryFrom<&BasicTypeEnum<'ctx>> for LLVMType {
 
 /// Conversion from Inkwell's array type to our type language.
 impl<'ctx> TryFrom<ArrayType<'ctx>> for LLVMType {
-    type Error = llvm_compile::Error;
+    type Error = llvm::Error;
 
     fn try_from(value: ArrayType<'ctx>) -> Result<Self, Self::Error> {
         Self::try_from(&value)
@@ -335,7 +335,7 @@ impl<'ctx> TryFrom<ArrayType<'ctx>> for LLVMType {
 
 /// Conversion from Inkwell's array type to our type language.
 impl<'ctx> TryFrom<&ArrayType<'ctx>> for LLVMType {
-    type Error = llvm_compile::Error;
+    type Error = llvm::Error;
 
     fn try_from(value: &ArrayType<'ctx>) -> Result<Self, Self::Error> {
         let length = value.len() as usize;
@@ -346,7 +346,7 @@ impl<'ctx> TryFrom<&ArrayType<'ctx>> for LLVMType {
 
 /// Conversion from Inkwell's generic float type to our specific float types.
 impl<'ctx> TryFrom<FloatType<'ctx>> for LLVMType {
-    type Error = llvm_compile::Error;
+    type Error = llvm::Error;
 
     fn try_from(value: FloatType<'ctx>) -> Result<Self, Self::Error> {
         Self::try_from(&value)
@@ -355,7 +355,7 @@ impl<'ctx> TryFrom<FloatType<'ctx>> for LLVMType {
 
 /// Conversion from Inkwell's generic float type to our specific float types.
 impl<'ctx> TryFrom<&FloatType<'ctx>> for LLVMType {
-    type Error = llvm_compile::Error;
+    type Error = llvm::Error;
 
     fn try_from(value: &FloatType<'ctx>) -> Result<Self, Self::Error> {
         #[allow(clippy::cast_possible_wrap)] // Our byte size should never be large enough
@@ -377,7 +377,7 @@ impl<'ctx> TryFrom<&FloatType<'ctx>> for LLVMType {
 /// Conversion from Inkwell's generic integer type to our specific integer
 /// types.
 impl<'ctx> TryFrom<IntType<'ctx>> for LLVMType {
-    type Error = llvm_compile::Error;
+    type Error = llvm::Error;
 
     fn try_from(value: IntType<'ctx>) -> Result<Self, Self::Error> {
         Self::try_from(&value)
@@ -387,7 +387,7 @@ impl<'ctx> TryFrom<IntType<'ctx>> for LLVMType {
 /// Conversion from Inkwell's generic integer type to our specific integer
 /// types.
 impl<'ctx> TryFrom<&IntType<'ctx>> for LLVMType {
-    type Error = llvm_compile::Error;
+    type Error = llvm::Error;
 
     fn try_from(value: &IntType<'ctx>) -> Result<Self, Self::Error> {
         let res = match value.get_bit_width() {
@@ -411,7 +411,7 @@ impl<'ctx> TryFrom<&IntType<'ctx>> for LLVMType {
 /// pointers. Otherwise, we would have to change every site performing
 /// conversion of pointer types.
 impl<'ctx> TryFrom<PointerType<'ctx>> for LLVMType {
-    type Error = llvm_compile::Error;
+    type Error = llvm::Error;
 
     fn try_from(value: PointerType<'ctx>) -> Result<Self, Self::Error> {
         Self::try_from(&value)
@@ -425,7 +425,7 @@ impl<'ctx> TryFrom<PointerType<'ctx>> for LLVMType {
 /// pointers. Otherwise, we would have to change every site performing
 /// conversion of pointer types.
 impl<'ctx> TryFrom<&PointerType<'ctx>> for LLVMType {
-    type Error = llvm_compile::Error;
+    type Error = llvm::Error;
 
     fn try_from(_: &PointerType<'ctx>) -> Result<Self, Self::Error> {
         Ok(Self::ptr)
@@ -434,7 +434,7 @@ impl<'ctx> TryFrom<&PointerType<'ctx>> for LLVMType {
 
 /// Conversion from Inkwell's struct type to our type language.
 impl<'ctx> TryFrom<StructType<'ctx>> for LLVMType {
-    type Error = llvm_compile::Error;
+    type Error = llvm::Error;
 
     fn try_from(value: StructType<'ctx>) -> Result<Self, Self::Error> {
         Self::try_from(&value)
@@ -443,7 +443,7 @@ impl<'ctx> TryFrom<StructType<'ctx>> for LLVMType {
 
 /// Conversion from Inkwell's struct type to our type language.
 impl<'ctx> TryFrom<&StructType<'ctx>> for LLVMType {
-    type Error = llvm_compile::Error;
+    type Error = llvm::Error;
 
     fn try_from(value: &StructType<'ctx>) -> Result<Self, Self::Error> {
         let field_types: Vec<Self> = value
@@ -463,7 +463,7 @@ impl<'ctx> TryFrom<&StructType<'ctx>> for LLVMType {
 /// seamlessly add support without having to change multiple conversion sites
 /// that would currently need to produce errors.
 impl<'ctx> TryFrom<VectorType<'ctx>> for LLVMType {
-    type Error = llvm_compile::Error;
+    type Error = llvm::Error;
 
     fn try_from(value: VectorType<'ctx>) -> Result<Self, Self::Error> {
         Self::try_from(&value)
@@ -477,7 +477,7 @@ impl<'ctx> TryFrom<VectorType<'ctx>> for LLVMType {
 /// seamlessly add support without having to change multiple conversion sites
 /// that would currently need to produce errors.
 impl<'ctx> TryFrom<&VectorType<'ctx>> for LLVMType {
-    type Error = llvm_compile::Error;
+    type Error = llvm::Error;
 
     fn try_from(value: &VectorType<'ctx>) -> Result<Self, Self::Error> {
         Err(Error::UnsupportedType(value.to_string()))?
@@ -486,7 +486,7 @@ impl<'ctx> TryFrom<&VectorType<'ctx>> for LLVMType {
 
 /// Conversion from Inkwell's function type to our type language.
 impl<'ctx> TryFrom<FunctionType<'ctx>> for LLVMType {
-    type Error = llvm_compile::Error;
+    type Error = llvm::Error;
 
     fn try_from(value: FunctionType<'ctx>) -> Result<Self, Self::Error> {
         Self::try_from(&value)
@@ -495,7 +495,7 @@ impl<'ctx> TryFrom<FunctionType<'ctx>> for LLVMType {
 
 /// Conversion from Inkwell's function type to our type language.
 impl<'ctx> TryFrom<&FunctionType<'ctx>> for LLVMType {
-    type Error = llvm_compile::Error;
+    type Error = llvm::Error;
 
     fn try_from(value: &FunctionType<'ctx>) -> Result<Self, Self::Error> {
         let return_type = value.get_return_type().map_or(Ok(LLVMType::void), Self::try_from)?;
@@ -514,7 +514,7 @@ impl<'ctx> TryFrom<&FunctionType<'ctx>> for LLVMType {
 /// We centralize this in a conversion to ensure that it is consistent at all
 /// use sites.
 impl<'ctx> TryFrom<VoidType<'ctx>> for LLVMType {
-    type Error = llvm_compile::Error;
+    type Error = llvm::Error;
 
     fn try_from(value: VoidType<'ctx>) -> Result<Self, Self::Error> {
         Self::try_from(&value)
@@ -526,7 +526,7 @@ impl<'ctx> TryFrom<VoidType<'ctx>> for LLVMType {
 /// We centralize this in a conversion to ensure that it is consistent at all
 /// use sites.
 impl<'ctx> TryFrom<&VoidType<'ctx>> for LLVMType {
-    type Error = llvm_compile::Error;
+    type Error = llvm::Error;
 
     fn try_from(_: &VoidType<'ctx>) -> Result<Self, Self::Error> {
         Ok(Self::void)
