+++ crates/compiler/src/pass/mod.rs
@@ -41,7 +41,7 @@ use std::{
 };
 
 use downcast_rs::Downcast;
-use hieratika_errors::llvm_compile::{Error, Result};
+use hieratika_errors::compile::llvm::{Error, Result};
 
 use crate::{
     context::SourceContext,
