+++ crates/error/src/compile/mod.rs
@@ -1,0 +1,2 @@
+pub mod cairo;
+pub mod llvm;
